pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "armymemo"
version = "0.1.0"
description = "Standalone Army memorandum library with Typst rendering"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Joe Schlessinger", email = "jschless@example.com" },
]
keywords = ["army", "memo", "typst", "correspondence", "pdf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Government",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "pdfplumber>=0.11.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
]

[project.scripts]
armymemo = "armymemo.cli:main"

[project.urls]
Repository = "https://github.com/jschless/armymemo"
Issues = "https://github.com/jschless/armymemo/issues"

[tool.setuptools.packages.find]
include = ["armymemo*"]
exclude = ["tests*", "docs*", "resources*"]

[tool.setuptools.package-data]
armymemo = [
    "resources/*.png",
    "resources/typst/*",
    "resources/rules/*",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
